Add get_config function
